+++ app/src/androidTest/java/com/duckduckgo/app/bookmarks/model/BookmarksDataRepositoryTest.kt
@@ -25,8 +25,6 @@ import com.duckduckgo.app.global.db.AppDatabase
 import com.duckduckgo.sync.store.Entity
 import com.duckduckgo.sync.store.EntityType.BOOKMARK
 import com.duckduckgo.sync.store.Relation
-import com.duckduckgo.sync.store.SyncEntitiesDao
-import com.duckduckgo.sync.store.SyncRelationsDao
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Assert.*
 import org.junit.Before
