--- app/src/main/java/com/duckduckgo/app/bookmarks/model/TreeNode.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2021 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.app.bookmarks.model
-
-class TreeNode<T>(val value: T) {
-
-    private val children: MutableList<TreeNode<T>> = mutableListOf()
-    fun add(child: TreeNode<T>) = children.add(child)
-    fun isEmpty() = children.isEmpty()
-
-    fun forEachVisit(
-        visitBefore: Visitor<T>,
-        visitAfter: Visitor<T>,
-    ) {
-        visitBefore(this)
-        children.forEach {
-            it.forEachVisit(visitBefore, visitAfter)
-        }
-        visitAfter(this)
-    }
-}
-
-typealias Visitor<T> = (TreeNode<T>) -> Unit
