--- sync/sync-api/src/main/java/com/duckduckgo/sync/api/Models.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.sync.api
-
-sealed class Entity(
-    open val id: String,
-    open val title: String,
-) {
-    data class Bookmark(
-        override val id: String,
-        override val title: String,
-        val url: String,
-    ) : Entity(id, title)
-
-    data class Favourite(
-        override val id: String,
-        override val title: String,
-        val url: String,
-    ) : Entity(id, title)
-
-    data class Folder(
-        override val id: String,
-        override val title: String,
-        val bookmarks: List<Bookmark>,
-    ) : Entity(id, title)
-}
