--- sync/sync-api/src/main/java/com/duckduckgo/sync/api/SyncRepository.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.sync.api
-
-import com.duckduckgo.sync.api.Entity.Bookmark
-import com.duckduckgo.sync.api.Entity.Favourite
-import com.duckduckgo.sync.api.Entity.Folder
-
-interface SyncRepository {
-
-    fun bookmarks(): List<Bookmark>
-    fun favorites(): List<Favourite>
-    fun allFolders(): List<Folder>
-
-    fun byFolder(folder: Folder): List<Folder>
-}
